src/requirements.txt
torch==2.0.1
torchvision==0.15.2
open_clip_torch==2.20.0
huggingface_hub==0.16.4
tokenizers==0.13.3
transformers>=4.31
opencv-python-headless==4.8.0.76
numpy==1.24.4
matplotlib==3.8
Pillow==10.0.0
scipy==1.12
timm==0.9.5
scikit-learn==1.3.2
nibabel